+++ rana_qgis_plugin/communication.py
@@ -3,6 +3,7 @@
 from qgis.core import Qgis, QgsMessageLog
 from qgis.PyQt.QtCore import Qt
 from qgis.PyQt.QtWidgets import QInputDialog, QMessageBox, QProgressBar, QPushButton
+from threedi_mi_utils.ui import ColoredProgressBar
 
 
 class UICommunication:
@@ -141,7 +142,7 @@ class UICommunication:
         if clear_msg_bar:
             self.iface.messageBar().clearWidgets()
         pmb = self.iface.messageBar().createMessage(msg)
-        pb = QProgressBar()
+        pb = ColoredProgressBar()
         pb.setMinimum(minimum)
         pb.setMaximum(maximum)
         pb.setValue(init_value)